Refactor honeypot into standalone lib
